chore: apply 0055, regen types, drop developer_id s8() shims
Migration 0055 applied to the Bazar project (ztxbguvmwpqccxuqwdqa);
properties.developer_id now exists (nullable uuid FK, backfilled from
development_id where present) and db/types.ts is regenerated against the
live schema. With developer_id in the generated types, the temporary
s8() casts on the property fetch (page.tsx) and the publish re-check
(_actions.ts loadPublishabilityFor) are no longer needed.

+++ app/(admin)/admin/properties/[id]/_actions.ts
@@ -677,9 +677,7 @@ async function loadPublishabilityFor(propertyId: string): Promise<{
   error: string | null;
 }> {
   const supabase = await createSupabaseServerClient();
-  // s8() widens the client so `developer_id` (migration 0055, not yet in
-  // db/types.ts) typechecks in the select + return.
-  const { data: p, error } = await s8(supabase)
+  const { data: p, error } = await supabase
     .from("properties")
     .select(
       "id, slug, reference, status, title, price_aed, developer_id, listing_permit_no, listing_permit_expires_at, compliance, property_media(role)",
